Move files from `pkg/sdk` to `meter`
Moved factory, factory_test, global, and option files from the `pkg/sdk` directory to the `meter` directory and updated their package names accordingly.
